pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sec-filing-rag"
version = "0.1.0"
description = "RAG application for SEC filings with page citations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "chromadb>=1.0.0",
  "openai>=1.30.0",
  "pymupdf>=1.24.0",
  "python-dotenv>=1.0.1",
  "streamlit>=1.44.0",
  "tiktoken>=0.9.0",
]

[project.optional-dependencies]
local-embeddings = [
  "sentence-transformers>=3.0.0",
]

[tool.setuptools.packages.find]
include = ["app*", "ingestion*", "retrieval*"]
